aaaaa

+++ database/migrations/2025_05_09_134257_create_servicos_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('servicos', function (Blueprint $table) {
             $table->id();
             $table->string('nome');
-            $table->decimal('valor', 8, 2);
+            $table->integer('valor');
             $table->timestamps();
         });
     }
